requirements.txt
dask>=2023.3.0
dask-ml>=2023.3.0
distributed>=2023.3.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0
bokeh>=2.4.0
plotly>=5.5.0
python-dotenv>=0.19.0
pytest>=6.2.0
black>=21.12b0
flake8>=4.0.0
isort>=5.10.0
pyyaml>=6.0
tqdm>=4.62.0
pillow>=9.0.0  # For image processing if needed
nltk>=3.6.0    # For text processing if needed
